imported maetIcon & renamed component files
