--- app/components/Autocomplete.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-type Props = {};
-
-export const Autocomplete = (props: Props) => <div>AutoSelect</div>;
