+++ stories/competition-card.tsx
